Add a VS Code setting to enable the experimental formatter (#273)
## Summary

This PR adds an `enableExperimentalFormatter` flag to the extension
which turns on formatting capabilities by setting the environment
variable in the LSP server. Our extension is setup such that if you
change this setting, we restart the server, so it is respected across
modifications. This is (intentionally) more discoverable, as I had
trouble with the environment variable alone.

## Test Plan

- Run the extension locally.
- Set `"editor.defaultFormatter": "charliermarsh.ruff"` in
`settings.json`.
- Try to format; verify that Ruff is not a valid option.
- Enable the setting (`"ruff.enableExperimentalFormatter": true`).
- Try to format; verify that it works as expected.
- Disable the setting; try to format; verify that it fails as expected:

<img width="1512" alt="Screen Shot 2023-09-02 at 5 23 25 PM"
src="https://github.com/astral-sh/ruff-vscode/assets/1309177/87fb4bb8-41f2-4454-9787-432123a8bddb">