Update README.md with example for using `ruff.args` in `settings.json` (#283)
Based on a suggestion in
https://github.com/astral-sh/ruff/discussions/7474#discussioncomment-7042823,
I'm making this change to make it easier for newbies like me to
customize ruff using settings.json.

Specifically using `ruff.args` in `settings.json`

<!--
Thank you for contributing to Ruff! To help us out with reviewing,
please consider the following:

- Does this pull request include a summary of the change? (See below.)
- Does this pull request include a descriptive title?
- Does this pull request include references to any relevant issues?
-->

## Summary

<!-- What's the purpose of the change? What does it do, and why? -->

As a newbie i didn't know i can use `ruff.args`. see
https://github.com/astral-sh/ruff/discussions/7474#discussioncomment-7042823.
So I added an example in the vscode extension readme.

## Test Plan

<!-- How was it tested? -->
This is README, so no tests needed

---------

Signed-off-by: KimSia, Sim <245021+simkimsia@users.noreply.github.com>
Co-authored-by: Charlie Marsh <charlie.r.marsh@gmail.com>