Support prioritizing discovered TOML settings over editor settings (#457)
## Summary

Fixes https://github.com/astral-sh/ruff-vscode/issues/425 (though only
in the pre-release version, since this is a feature for the new LSP
server)

This is a follow-up to
https://github.com/astral-sh/ruff-vscode/pull/456.

A new setting has been introduced, `Prioritize File Configuration`,
which tells the extension to prioritize settings from discovered TOML
files over extension settings. Extension settings will still be used
when a set field in the extension is not set in the file configuration.

## Test Plan

Refer to https://github.com/astral-sh/ruff/pull/11086 for a test plan.