Avoid querying the Python extension for the interpreter (#528)
## Summary

Fix the CI failure on `main`. I think the tests are currently
unpredictable in the sense that the tests could begin _before_ the
server starts which then creates the failure. The main reason I believe
this is happening is because the extension uses a PubSub mechanism here:


https://github.com/astral-sh/ruff-vscode/blob/80d1b13a80ac99afb96a34ae6358d893158a65b8/src/common/python.ts#L250-L259

This change is so that it never hits this by always testing with the
bundled executable and always using the native server.

## Test Plan

I don't really have a good way to test this until it's being triggered
multiple times which will verify the consistency now.