Add `logLevel` and `logFile` extension setting (#530)
## Summary

This PR adds `ruff.logLevel` and `ruff.logFile` settings to the VS Code
extension which corresponds to the same for the `ruff server`. Without
these, there's no way to configure the log level and the destination for
the log messages unlike other editors.

## Test Plan

Using the following config:
```json
{
  "ruff.logLevel": "debug",
  "ruff.logFile": "~/playground/ruff/server.logs",
  "ruff.trace.server": "messages"
}
```

I confirmed that it outputs the debug logs messages at the specified location.