Update to reflect the stabilized `ruff server` (#531)
## Summary

This PR updates the README to reflect the stabilized `ruff server`. It
also changes the minimum required version for the usage of `ruff server`
in `nativeServer: auto` to be `0.5.3`.

For the README, the settings that are just a _pass through_ to the
native server are removed and instead we point the users to the online
settings reference. Now, the settings table only contains the ones which
are specific to VS Code extension.

For review, here's the rendered README:
https://github.com/astral-sh/ruff-vscode/blob/dhruv/readme/README.md

resolves: #520