Use `spawn` rather than `exec` to support paths with spaces (#539)
## Summary

If the interpreter path contains a space, then the find script fails. We
need to escape the arguments that we pass to `exec`. But, it seems
easier to just use `spawn`, which doesn't require escaping.

Closes https://github.com/astral-sh/ruff/issues/12394.

## Test Plan

Created a virtual environment `foo bar` in a repo; verified that the
script failed before but succeeds after this change.