Use extension context for output window disposable (#671)
## Summary

I noticed this in https://github.com/astral-sh/ruff-vscode/pull/585 that
we would not be able to look at the error message when the server
crashes because the output window got disposed. This is because the
disposable is tied up with stopping the server. This PR updates that to
include the disposable to the extension context so that it gets disposed
only when the extension is deactivated.

## Test Plan

Run the scenario from https://github.com/astral-sh/ruff-vscode/pull/585
to make sure that the "Ruff Language Server" window is not disposed in
case the server crashed.