Allow arbitrary config keys in `ruff.configuration` (#702)
## Summary

Closes: #6 

This PR adds support for https://github.com/astral-sh/ruff/pull/16296
for the VS Code extension by allowing any arbitrary object in
`ruff.configuration`

Additionally, it will provide a warning message to the user if they're
using inline configuration but the Ruff version does not support it.

It has one disadvantage that the user will see two popups where the
error one is coming from the server because it cannot deserialize the
options. We could not show the warning popup if this is too much. I'm
worried that it might go unnoticed because the "Show Logs" in the below
popup will open the server logs while the message would be in the client
logs.

<img width="491" alt="Screenshot 2025-02-25 at 1 24 15 PM"
src="https://github.com/user-attachments/assets/8bafbd69-f8fa-4604-8ab3-1f6efa745045"
/>

We'll still log it on the console (client log channel):
```
2025-02-25 13:24:10.067 [warning] Inline configuration support was added in Ruff 0.9.8 (current version is 0.9.7). Please update your Ruff version to use this feature.
```

I haven't provided more details in the warning message based on the
assumption that if the user is using inline configuration then they're
aware of it but it's just that the Ruff version is too old.

## Test Plan

Refer to the test plan in https://github.com/astral-sh/ruff/pull/16296