Add some more assertions to regexes in the release script (#729)
## Summary

This PR adds some assertions to the release script that makes sure that:
- There is exactly one match of the description regex in the README
- There is at least one match of the svg regex in the README

This makes the script more robust to unexpected changes in Ruff's
versioning scheme or the prose in the README.

## Test Plan

- I ran the script on this branch as-is and it ran successfully without
updating the README (because there's currently nothing to update)
- I ran `git checkout 5ea1add14470e3bb15b8bfd3580c5fddeb3d117f --
README.md CHANGELOG.md package.json package-lock.json pyproject.toml
requirements.txt requirements-dev.txt` to revert all changes from the
last release. I then re-ran the release script. The diff relative to
this branch was:

  ```diff
  diff --git a/CHANGELOG.md b/CHANGELOG.md
  index 3f3250c..5185006 100644
  --- a/CHANGELOG.md
  +++ b/CHANGELOG.md
@@ -4,7 +4,7 @@ See [here](https://github.com/astral-sh/ruff/releases)
for the Ruff release note
 
   ## 2025.20.0
 
-This release upgrades the bundled Ruff version to `v0.11.1`. This
includes a bug fix to allow `Fix All` and `Organize Imports` actions in
the presence of version-specific syntax errors.
  +This release upgrades the bundled Ruff version to `v0.11.1`.
  ```

which indicates that it also still runs successfully even when there are
updates in the README to make.