Bump Ubuntu version in release workflow (#731)
## Summary

I noticed the warnings about the Ubuntu 20.04 brownout during the
release today. We're actually only 10 days from the end of the brownout
period and when 20.04 is fully unsupported
([ref](https://github.com/actions/runner-images/issues/11101)).

There's one additional mention of 20.04 a bit farther down in the file
for the `run-on-arch` action, but that's the same with Ruff.

Possibly we could also bump all the way to 24.04 if we wanted.